+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 Notable changes to this project.
 
+## dev
+- support passing `round_num` to `list_datasets`, to get available files from past rounds
+
 ## [2.8.1] - 2021-09-08
 -  Add version arg to upload_predictions (#59)