+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 Notable changes to this project.
 
-## dev
+## [2.9.0] - 2021-09-15
 - support passing `round_num` to `list_datasets`, to get available files from past rounds
 - `download_dataset` no longer requires a destination path, it defaults to the source file name
 - `download_dataset` now accepts a `round_num` argument, to download old files